Split display domains: sprite_animation + terrain_animation; slim map to static tilemap
- animation/ renamed sprite_animation/ (frame-table creature animation;
  the name no longer overclaims 'animation')
- map/ split: terrain scroll spec/layers/system move to new
  terrain_animation/ domain (UV-scroll mechanism, coupled to chunks'
  transparent shoreline tiles); TerrainAnimSpec value type lands in
  types/, constants keeps only the TERRAIN_ANIMS table
- map/ becomes the static tilemap domain: chunks, layout, coords,
  autotile (stitching renamed; PD's getWaterTile is 4-bit bitmask
  autotiling)
- chunks.rs: extract pure build_chunk_data() (GridMap -> tile arrays)
  into utils/ for testability; TILE_SIZE i32 -> f32 (display pixel math
  is floating point)
- camera absorbs the hero glue as attach_target in DisplayPhase::Attach;
  display/hero/ domain deleted
- map domain stabilized: cleanup TODO headers dropped

+++ src/frontend/display/map/constants/layout.rs
@@ -1,9 +1,10 @@
-// TODO: pending cleanup review — remove once stabilized
 //! Display-side layout constants: pixel size, z layers, tileset indices.
 //! Map dimensions in cells are game data (`core::map::constants::layout`).
 
-/// Tile size in world pixels; every rendered grid constant derives from it.
-pub const TILE_SIZE: i32 = 16;
+/// Tile size in world pixels; every rendered grid constant derives from
+/// it. `f32`: the display side's pixel math is floating point (the rare
+/// integer use, e.g. texture sizes, casts at the call site).
+pub const TILE_SIZE: f32 = 16.0;
 
 // Fixed z layers. The scrolling terrain layer sits beneath the floor and
 // shows through the chunks' empty cells; actors render between floor and
