+++ src/frontend/display/map/utils/coords.rs
@@ -1,4 +1,3 @@
-// TODO: pending cleanup review — remove once stabilized
 //! Cell <-> world-pixel conversion. The core counts in continuous cell
 //! coordinates and knows nothing about pixels; these functions are the
 //! display side's only bridge between the two spaces.
@@ -13,17 +12,14 @@ use crate::frontend::display::map::constants::layout::TILE_SIZE;
 /// are cell centers, hence the half-cell offset. Map row 0 is the top row
 /// while world Y points up, so world y is negative.
 pub fn cell_to_world(pos: Position) -> Vec2 {
-    Vec2::new(
-        (pos.x + 0.5) * TILE_SIZE as f32,
-        -(pos.y + 0.5) * TILE_SIZE as f32,
-    )
+    Vec2::new((pos.x + 0.5) * TILE_SIZE, -(pos.y + 0.5) * TILE_SIZE)
 }
 
 /// Cell containing a world-pixel position.
 pub fn world_to_cell(world: Vec2) -> CellCoord {
     CellCoord::new(
-        (world.x / TILE_SIZE as f32).floor() as i32,
-        (-world.y / TILE_SIZE as f32).floor() as i32,
+        (world.x / TILE_SIZE).floor() as i32,
+        (-world.y / TILE_SIZE).floor() as i32,
     )
 }
